# 2NotyBackend/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "2Noty services",
        "description": "Here!, you'll can explorer all services like to model, response and more!, good Look!",
        "contact": {
            "name": "Hugo Ivan Cruz Hernández",
            "email": "ing.hugo.isc@gmail.com.com"
        },
        "license": {
            "name": "2Noty",
            "url": ""
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "paths": {
...
        "/api/subscriptions/updateSubscription": {
            "put": {
                "operationId": "updateSubscription",
                "tags": ["Subscriptions"],
                "summary": "update a subscription",
                "description": "Update a subscription",
                "consumes": "application/json",
                "parameters": [{
                    "in": "body",
                    "name": "2Noty",
                    "description": "object to send from client",
                    "schema": {
                        "$ref": "#/definitions/updateSubscription"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Data",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "ok": {
                                    "type": "string",
                                    "example": "ok"
                                },
                                "data": {
                                    "type": "string",
                                    "example": "update"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "updateSubscription": {
            "type": "object",
            "properties": {
                "id_suscripcion": {
                    "type": "integer",
                    "example": 1
                },
                "id_pais": {
                    "type": "integer",
                    "example": 1
                },
                "id_empresa": {
                    "type": "integer",
                    "example": 1
                },
                "id_marca": {
                    "type": "integer",
                    "example": 1
                },
                "id_categoria_suscripcion": {
                    "type": "integer",
                    "example": 1
                },
                "suscripcion": {
                    "type": "string",
                    "example": "Prueba"
                },
                "descripcion": {
                    "type": "string",
                    "example": "Prueba"
                },
                "id_estatus": {
                    "type": "integer",
                    "example": 1
                },
                "uel_imagen": {
                    "type": "string",
                    "example": "http://example"
                },
                "url_icono": {
                    "type": "string",
                    "example": "http://example"
                }
            }
        },
...
    },
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"]
}